:art: Fixed some bugs in node attachment to the master and the deletion of the node from cluster.
there was some piece of code was missing like while deleting the node we are not deleting from the cluster. so added that part so now it will delete the node from the kubernetes cluster also. and while creating the cluster it will not throw error if the cluster already created it will just restart that existing cluster.